requirements.txt
absl-py==2.3.1
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.6.1
appnope==0.1.3
archspec @ file:///home/conda/feedstock_root/build_artifacts/archspec_1737352602016/work
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
backports.tarfile==1.2.0
backports.zstd==1.1.0
beautifulsoup4==4.11.1
black==25.12.0
bleach==5.0.1
boltons @ file:///home/conda/feedstock_root/build_artifacts/boltons_1733827268945/work
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1725267563793/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1739515848642/work/certifi
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1725560574711/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1735929714516/work
click==8.1.3
cloudpickle==2.2.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
coloredlogs==15.0.1
conda @ file:///Users/runner/miniforge3/conda-bld/conda_1698450890070/work
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1736345463896/work
conda_package_streaming @ file:///home/conda/feedstock_root/build_artifacts/conda-package-streaming_1741620732069/work
contourpy==1.0.5
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography-split_1740893596609/work
cycler==0.11.0
dask==2022.9.2
dask-jobqueue==0.8.1
datar==0.12.2
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
-e git+https://github.com/facebookresearch/detectron2.git@fd27788985af0f4ca800bca563acdb700bb890e2#egg=detectron2
diot==0.2.2
distlib==0.4.0
distributed==2022.9.2
easyocr==1.7.2
einops==0.8.2
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.37.4
fsspec==2026.2.0
ftfy==6.1.1
fvcore==0.1.5.post20221221
gast==0.7.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
h5py==3.15.1
hatch==1.16.1
hatchling==1.28.0
HeapDict==1.0.1
hf-xet==1.2.0
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
humanfriendly==10.0
hydra-core==1.3.2
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
hyperlink==21.0.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
imageio==2.22.1
imgviz==1.7.6
importlib_metadata==8.7.0
imutils==0.5.4
inflection==0.5.1
iopath==0.1.10
ipykernel==6.16.0
ipympl==0.9.8
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.10
jsonpatch @ file:///home/conda/feedstock_root/build_artifacts/jsonpatch_1733814567314/work
jsonpointer @ file:///Users/runner/miniforge3/conda-bld/jsonpointer_1725302947603/work
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-server==1.19.1
jupyter_client==7.3.5
jupyterlab==3.4.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.15.2
kaggle==1.7.4.5
kagglehub==0.3.13
keras==3.12.0
keyring==25.7.0
kiwisolver==1.4.4
labelme==5.10.1
libclang==18.1.1
locket==1.0.0
loguru==0.7.3
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.1
matplotlib==3.10.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
mizani==0.8.1
ml-dtypes==0.3.2
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.0.4
mypy_extensions==1.1.0
namex==0.1.0
narwhals==1.33.0
natsort==8.4.0
nbclassic==0.4.4
nbclient==0.6.8
nbconvert==7.1.0
nbformat==5.6.1
nest-asyncio==1.5.6
networkx==2.8.7
ninja==1.13.0
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.23.2
opencv-contrib-python==4.6.0.66
opencv-python==4.11.0.86
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optree==0.18.0
osam==0.2.5
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
palettable==3.3.0
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.3.0
pathspec==0.12.1
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pipda==0.12.0
plantcv==4.7
platformdirs==4.5.0
plotnine==0.10.1
pluggy @ file:///home/conda/feedstock_root/build_artifacts/pluggy_1733222765875/work
portalocker==3.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==4.25.8
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyclipper==1.3.0.post6
pycocotools==2.0.10
pycosat @ file:///Users/runner/miniforge3/conda-bld/pycosat_1732588454015/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.13.0
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1737243356468/work
pyparsing==3.0.9
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5_sip==12.17.1
pyrsistent==0.18.1
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
pytesseract @ git+https://github.com/madmaze/pytesseract.git@df9fce0b5c5ea3cd9182fce8870ab9241d951173
python-bidi==0.6.7
python-dateutil==2.8.2
python-simpleconf==0.6.0
python-slugify==8.0.4
pytokens==0.3.0
pytz==2022.4
PyWavelets==1.4.1
PyYAML==5.1
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
regex==2026.1.15
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1733217035951/work
rich==14.2.0
ruamel.yaml @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml_1698138674285/work
ruamel.yaml.clib @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml.clib_1728724504331/work
safetensors==0.7.0
# Editable install with no version control (sam3==0.1.0)
-e /Users/quino070/sam3
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.15.3
Send2Trash==1.8.0
shapely==2.1.2
shellingham==1.5.4
simplug==0.3.2
six==1.17.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spectral==0.23.1
stack-data==0.5.1
statsmodels==0.13.2
sympy==1.14.0
tabulate==0.9.0
tblib==1.7.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.2.0
terminado==0.16.0
text-unidecode==1.3
tf_keras==2.16.0
threadpoolctl==3.1.0
tifffile==2022.8.12
timm==1.0.24
tinycss2==1.1.1
tokenizers==0.22.2
tomli==2.0.1
tomli_w==1.2.0
tomlkit==0.13.3
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tornado==6.4.2
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
traitlets==5.4.0
transformers==5.1.0
trove-classifiers==2025.12.1.14
truststore @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_truststore_1739009763/work
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1734859416348/work
userpath==1.9.2
utils==1.0.1
uv==0.9.15
virtualenv==20.35.4
vl-convert-python==1.7.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==3.1.4
widgetsnbextension==4.0.3
wrapt==2.0.1
xarray==2025.3.1
yacs==0.1.8
zict==2.2.0
zipp==3.23.0
zstandard==0.23.0
